backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "serverless-financial-pipeline-backend"
version = "0.1.0"
description = "AWS Lambda backend for the serverless financial pipeline"
requires-python = ">=3.10"
dependencies = [
  "boto3>=1.35,<2",
  "requests>=2.31,<3",
  "pandas>=2.0,<3",
  "numpy>=1.25,<3",
  "python-dotenv>=1.0,<2",
]

[tool.uv]
# Use uv for dependency management (uv sync / uv run)
dev-dependencies = [
  # "pytest>=7.0,<9",
  # "ruff>=0.5.0,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["serverless_financial_pipeline_backend"]


